smart_orchard/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
openai==1.12.0
scikit-learn==1.4.0
numpy==1.26.0
requests==2.31.0
Werkzeug==3.0.1
cryptography==42.0.0